requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
pandas==2.0.3
plotly==5.17.0
openpyxl==3.1.2
weasyprint==60.2